+++ packages/seroval/src/core/cross/serializer.ts
@@ -24,10 +24,7 @@ export default class CrossSerializerContext extends BaseSerializerContext {
     this.scopeId = options.scopeId;
   }
 
-  getRefParam(id: number | string): string {
-    if (typeof id === 'string') {
-      return GLOBAL_CONTEXT_REFERENCES + '.' + id;
-    }
+  getRefParam(id: number): string {
     return GLOBAL_CONTEXT_REFERENCES + '[' + id + ']';
   }
 
@@ -35,24 +32,37 @@ export default class CrossSerializerContext extends BaseSerializerContext {
     index: number,
     value: string,
   ): string {
+    // In cross-reference, we have to assume that
+    // every reference are going to be referenced
+    // in the future, and so we need to store
+    // all of it into the reference array.
     return this.getRefParam(index) + '=' + value;
   }
 
   serializeTop(tree: SerovalNode): string {
+    // Get the serialized result
     const result = this.serialize(tree);
+    // If the node is a non-reference, return
+    // the result immediately
     const id = tree.i;
     if (id == null) {
       return result;
     }
+    // Get the patches
     const patches = this.resolvePatches();
+    // Get the variable that represents the root
     const ref = this.getRefParam(id);
+    // Parameters needed for scoping
     const params = this.scopeId == null ? '' : GLOBAL_CONTEXT_REFERENCES;
-    const mainBody = patches ? result + ',' + patches : result;
+    // If there are patches, append it after the result
+    const body = patches ? result + ',' + patches + ref : result;
+    // If there are no params, there's no need to generate a function
     if (params === '') {
-      return patches ? '(' + mainBody + ref + ')' : mainBody;
+      return patches ? '(' + body + ')' : body;
     }
+    // Get the arguments for the IIFE
     const args = this.scopeId == null ? '()' : '(' + GLOBAL_CONTEXT_REFERENCES + '["' + serializeString(this.scopeId) + '"])';
-    const body = mainBody + (patches ? ref : '');
+    // Create the IIFE
     return '(' + this.createFunction([params], body) + ')' + args;
   }
 }
