+++ packages/seroval/src/core/tree/serializer.ts
@@ -19,7 +19,7 @@ export default class VanillaSerializerContext extends BaseSerializerContext {
    * Map tree refs to actual refs
    * @private
    */
-  valid = new Map<number | string, number>();
+  valid = new Map<number, number>();
 
   /**
    * Variables
@@ -31,7 +31,7 @@ export default class VanillaSerializerContext extends BaseSerializerContext {
    * Creates the reference param (identifier) from the given reference ID
    * Calling this function means the value has been referenced somewhere
    */
-  getRefParam(index: number | string): string {
+  getRefParam(index: number): string {
     /**
      * Creates a new reference ID from a given reference ID
      * This new reference ID means that the reference itself
